requirements.txt
pandas>=1.4.3,<2.0.0
requests
python-dotenv
dataset
pandera
pytest
pre-commit
polars-lts-cpu
typing_extensions
dateparser
litellm<1.40.0
pydantic<2.0.0
parsee-core<0.1.5.6
